+++ includes/pis-functions-display.php
@@ -1259,20 +1259,12 @@ function pis_debug( $args ) {
 			) . "\n";
 			$output    .= '<ul class="pis-debug-ul">' . "\n";
 			$output    .= '<li class="pis-debug-li">' . sprintf(
-				// translators: %s is the time when the cache was created.
+				// translators: %s is the date and time when the cache was created.
 				esc_html__(
 					'Created on: %s',
 					'posts-in-sidebar'
 				),
 				$cache_info['cache_created'] . '</li>' . "\n"
-			);
-			$output .= '<li class="pis-debug-li">' . sprintf(
-				// translators: %s is the duration of the cache.
-				esc_html__(
-					'Duration: %s',
-					'posts-in-sidebar'
-				),
-				$cache_info['cache_duration'] . '</li>' . "\n"
 			);
 			$output .= '<li class="pis-debug-li">' . sprintf(
 				// translators: %s is the time when the cache will expire.
@@ -1283,9 +1275,25 @@ function pis_debug( $args ) {
 				$cache_info['cache_expires'] . '</li>' . "\n"
 			);
 			$output .= '<li class="pis-debug-li">' . sprintf(
+				// translators: %s is the duration of the cache.
+				esc_html__(
+					'Duration: %s',
+					'posts-in-sidebar'
+				),
+				$cache_info['cache_duration'] . '</li>' . "\n"
+			);
+			$output    .= '<li class="pis-debug-li">' . sprintf(
+				// translators: %s is the time passed from cache creation.
+				esc_html__(
+					'Time passed: %s',
+					'posts-in-sidebar'
+				),
+				$cache_info['cache_passed'] . '</li>' . "\n"
+			);
+			$output .= '<li class="pis-debug-li">' . sprintf(
 				// translators: %s is the remaining time.
 				esc_html__(
-					'Remaining time: %s',
+					'Expires in: %s',
 					'posts-in-sidebar'
 				),
 				$cache_info['cache_remaining_time'] . '</li>' . "\n"